assorted fixes to ensure startup
