+++ modules/toolkit.py
@@ -1,6 +1,4 @@
 ## built-in modules
-from datetime import datetime
-
 import os
 import msvcrt
 import time
@@ -323,24 +321,6 @@ class Toolkit():
 
             return is_connection, update_prompt
     
-##-------------------start-of-get_timestamp()---------------------------------------------------------------------------------------------------------------------------------------------------------------------------
-
-    @staticmethod
-    def get_timestamp() -> str:
-
-        """
-        
-        Generates a timestamp for an action taken by Kudasai.
-
-        Returns:
-        time_stamp (string) : The timestamp for the action.        
-        
-        """
-
-        time_stamp = "[" + datetime.now().strftime("%Y-%m-%d %H:%M:%S") + "] "
-
-        return time_stamp
-
 ##--------------------start-of-UserCancelError------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------
 
     class UserCancelError(Exception):
